API - TASK - update task service

+++ src/modules/tasks/tasks.service.ts
@@ -1,7 +1,10 @@
+import { assertProjectAccess } from "../../lib/permissions";
 import { prisma } from "../../lib/prisma";
 import { CreateTaskBody, UpdateTaskBody } from "./tasks.types";
 
-export async function getAllTasks(projectId: string) {
+export async function getAllTasks(projectId: string, userId: string) {
+  await assertProjectAccess(prisma, projectId, userId);
+
   const tasks = await prisma.task.findMany({
     where: { projectId },
     select: {
@@ -21,7 +24,13 @@ export async function getAllTasks(projectId: string) {
   return tasks;
 }
 
-export async function getTaskById(id: string, projectId: string) {
+export async function getTaskById(
+  id: string,
+  projectId: string,
+  userId: string,
+) {
+  await assertProjectAccess(prisma, projectId, userId);
+
   const task = await prisma.task.findFirst({
     where: {
       id,
@@ -32,7 +41,13 @@ export async function getTaskById(id: string, projectId: string) {
   return task;
 }
 
-export async function createTask(input: CreateTaskBody, projectId: string) {
+export async function createTask(
+  input: CreateTaskBody,
+  projectId: string,
+  userId: string,
+) {
+  await assertProjectAccess(prisma, projectId, userId);
+
   const task = await prisma.task.create({
     data: {
       title: input.title,
@@ -50,7 +65,29 @@ export async function updateById(
   id: string,
   updates: UpdateTaskBody,
   projectId: string,
+  userId: string,
 ) {
+  const userMember = await assertProjectAccess(prisma, projectId, userId);
+
+  const task = await prisma.task.findFirst({
+    where: {
+      id,
+      projectId,
+    },
+    select: {
+      assigneeId: true,
+    },
+  });
+
+  const isAdmin = userMember.role === "ADMIN";
+  const isAssignee = task?.assigneeId === userId;
+
+  if (!isAdmin && !isAssignee) {
+    const error = new Error("FORBIDDEN");
+    (error as any).statusCode = 403;
+    throw error;
+  }
+
   const updatedTask = await prisma.$transaction(async (tx) => {
     const result = await tx.task.updateMany({
       where: { id, projectId },
@@ -75,7 +112,20 @@ export async function updateById(
   return updatedTask;
 }
 
-export async function deleteTaskById(id: string, projectId: string) {
+export async function deleteTaskById(
+  id: string,
+  projectId: string,
+  userId: string,
+) {
+  const userMember = await assertProjectAccess(prisma, projectId, userId);
+
+  if (userMember.role !== "ADMIN") {
+    const error = new Error("FORBIDDEN");
+    (error as any).statusCode = 403;
+
+    throw error;
+  }
+
   const deletedTask = await prisma.task.deleteMany({
     where: {
       id,
